Extracts rewriteCSSURLs private function into middleware-ish bits.
Additionally, the middleware is only used in the combine middleware if a basePath config is provided.
