+++ lib/combohandler.js
@@ -1,6 +1,5 @@
 var fs   = require('fs'),
     path = require('path'),
-    URI  = require("URIjs"),
 
     // exported to allow instanceof checks
     BadRequest = exports.BadRequest = require('./error/bad-request'),
@@ -21,6 +20,7 @@ exports.combine = function (config) {
     var maxAge    = config.maxAge,
         mimeTypes = config.mimeTypes || MIME_TYPES,
         basePath  = config.basePath || "",
+        rewriteCSSURLs,
 
         // Intentionally using the sync method because this only runs when the
         // middleware is initialized, and we want it to throw if there's an
@@ -29,6 +29,10 @@ exports.combine = function (config) {
 
     if (typeof maxAge === 'undefined') {
         maxAge = 31536000; // one year in seconds
+    }
+
+    if (basePath) {
+        rewriteCSSURLs = require('./middleware/cssUrls').rewrite;
     }
 
     return function combineMiddleware(req, res, next) {
@@ -182,20 +186,3 @@ function parseQuery(url) {
     return parsed;
 }
 
-function rewriteCSSURLs(basePath, relativePath, data) {
-    return data.replace(/[\s:]url\(\s*(['"]?)(\S+)\1\s*\)/g, function (substr, quote, match) {
-        // There is a ton of complexity related to URL parsing related
-        // to unicode, escapement, etc. Rather than try to capture that,
-        // this just does a simple sniff to validate whether the URL
-        // needs to be rewritten.
-        if (!URI(match).is("relative")) {
-            return substr;
-        }
-
-        var fileBasePath = path.join(basePath, relativePath),
-            fileDirName  = path.dirname(fileBasePath),
-            absolutePath = path.resolve(fileDirName, match);
-
-        return substr.replace(match, absolutePath);
-    });
-}
